+++ CHANGELOG.md
@@ -10,11 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `release.sh` quality of life automation script for releasing
+- param `simulationSpeed` (default 3) in `controllerOptions`: sets the initial simulation speed when simulation mode is entered
+- param `simulationAutoPlay` (default false) in `controllerOptions` allows to trigger default play as soon as simulation mode is entered
+- `BACKLOG.md` to group future evolutions demands
 
 ### Changed
 
 - Updated `README.md` with the **Use Cases** section (clarifications on the project's scope and simulation model capabilities)
 - Cleaned and updated `CLAUDE.md`
+- Set camera default max distance from 100 to 200
+- Upon initialization Engine emits a first `modeChanged` event. Therefore `previousMode` is now optional in that event.
 
 ### Removed